feat: remove contract upload flow and logic - Remove complete-application and upload contract endpoints and views - Remove CONTRACT_UPLOADED and COMPLETED status and methods, including from ToU - Add direct "Decide" button in application list for easy access - Add back and next buttons in all stages - Add icons to reject and approve buttons - Add amount disbursed field to stage five - Remove unused statistics from OCP and Lender dashboards - Remove SIGNED_CONTRACT special logic
