+++ src/api/private.ts
@@ -1,7 +1,6 @@
 /* eslint-disable camelcase */
 import {
   ApproveApplicationInput,
-  CompleteApplicationInput,
   EmailToSMEInput,
   IApplication,
   IApplicationsListResponse,
@@ -166,12 +165,6 @@ export const approveApplicationFn = async (approvePayload: ApproveApplicationInp
 export const rejectApplicationFn = async (rejectPayload: RejectApplicationInput) => {
   const { application_id, ...payload } = rejectPayload;
   const response = await authApi.post<IApplication>(`applications/${application_id}/reject-application`, payload);
-  return response.data;
-};
-
-export const completeApplicationFn = async (completePayload: CompleteApplicationInput) => {
-  const { application_id, ...payload } = completePayload;
-  const response = await authApi.post<IApplication>(`applications/${application_id}/complete-application`, payload);
   return response.data;
 };
