+++ src/api/public.ts
@@ -2,7 +2,6 @@ import {
   ApplicationBaseInput,
   ChangeEmailInput,
   ConfirmChangeEmailInput,
-  ContractAmountInput,
   DeclineApplicationInput,
   DeclineFeedbackInput,
   GetCreditProductsOptionsInput,
@@ -11,7 +10,6 @@ import {
   IBorrowerDocument,
   ICreditProduct,
   SelectCreditProductInput,
-  UploadContractInput,
   UploadFileInput,
 } from '../schemas/application';
 import { publicApi } from './axios';
@@ -83,16 +81,6 @@ export const applicationSubmitFn = async (payload: ApplicationBaseInput) => {
 
 export const aditionalDataSubmitFn = async (payload: ApplicationBaseInput) => {
   const response = await publicApi.post<IApplicationResponse>('applications/complete-information-request', payload);
-  return response.data;
-};
-
-export const uploadContractFn = async (payload: UploadContractInput) => {
-  const response = await publicApi.postForm<IBorrowerDocument>('applications/upload-contract', payload);
-  return response.data;
-};
-
-export const uploadContractConfirmationFn = async (payload: ContractAmountInput) => {
-  const response = await publicApi.post<IApplicationResponse>('applications/confirm-upload-contract', payload);
   return response.data;
 };
