+++ src/constants/index.ts
@@ -60,18 +60,11 @@ export enum APPLICATION_STATUS {
   SUBMITTED = 'SUBMITTED',
   STARTED = 'STARTED',
   APPROVED = 'APPROVED',
-  CONTRACT_UPLOADED = 'CONTRACT_UPLOADED',
-  COMPLETED = 'COMPLETED',
   REJECTED = 'REJECTED',
   INFORMATION_REQUESTED = 'INFORMATION_REQUESTED',
 }
 
-export const COMPLETED_STATUS = [
-  APPLICATION_STATUS.COMPLETED,
-  APPLICATION_STATUS.REJECTED,
-  APPLICATION_STATUS.APPROVED,
-  APPLICATION_STATUS.LAPSED,
-];
+export const COMPLETED_STATUS = [APPLICATION_STATUS.REJECTED, APPLICATION_STATUS.APPROVED, APPLICATION_STATUS.LAPSED];
 
 export const NOT_STARTED_STATUS = [APPLICATION_STATUS.SUBMITTED];
 
@@ -85,8 +78,6 @@ export const APPLICATION_STATUS_NAMES: { [key: string]: string } = {
   [APPLICATION_STATUS.SUBMITTED]: t('Submitted'),
   [APPLICATION_STATUS.STARTED]: t('Started'),
   [APPLICATION_STATUS.APPROVED]: t('Approved'),
-  [APPLICATION_STATUS.CONTRACT_UPLOADED]: t('Contract uploaded'),
-  [APPLICATION_STATUS.COMPLETED]: t('Completed'),
   [APPLICATION_STATUS.REJECTED]: t('Rejected'),
   [APPLICATION_STATUS.INFORMATION_REQUESTED]: t('Information requested'),
 };
@@ -143,8 +134,6 @@ export const STATISTICS_DATE_FILTER_OPTIONS: FormSelectOption[] = [
     label: t('Last month'),
   },
 ];
-
-export const SIGNED_CONTRACT_DOCUMENT_TYPE = 'SIGNED_CONTRACT';
 
 export const DEFAULT_BORROWER_SIZE = 'NOT_INFORMED';
