+++ src/layout/PublicApplicationLayout.tsx
@@ -33,20 +33,10 @@ export default function PublicApplicationLayout() {
       const { pathname } = location;
       const lastSegment = pathname.substring(pathname.lastIndexOf('/') + 1);
       if (lastSegment === 'change-primary-email') return;
-      if (application.status === APPLICATION_STATUS.COMPLETED) {
+      if (application.status === APPLICATION_STATUS.APPROVED) {
         if (lastSegment !== 'application-completed') navigate('./application-completed');
       } else if (application.status === APPLICATION_STATUS.REJECTED) {
         if (lastSegment !== 'find-alternative-credit' && lastSegment !== 'rejected') navigate('./rejected');
-      } else if (application.status === APPLICATION_STATUS.CONTRACT_UPLOADED) {
-        if (lastSegment !== 'upload-contract-completed') {
-          navigate('./upload-contract-completed');
-        }
-      } else if (application.status === APPLICATION_STATUS.APPROVED) {
-        if (lastSegment !== 'upload-contract' && !application.contract_amount_submitted) {
-          navigate('./upload-contract');
-        } else if (lastSegment !== 'upload-contract-completed' && application.contract_amount_submitted) {
-          navigate('./upload-contract-completed');
-        }
       } else if (
         application.status === APPLICATION_STATUS.SUBMITTED ||
         application.status === APPLICATION_STATUS.STARTED
