+++ src/pages/fi/StageTwo.tsx
@@ -4,7 +4,7 @@ import { Box, Dialog } from '@mui/material';
 import { useT } from '@transifex/react';
 import { useEffect, useState } from 'react';
 import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
-import { useNavigate } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import EmailIcon from 'src/assets/icons/email.svg';
 import StepImageEN from 'src/assets/pages/en/stage-two.svg';
 import StepImageES from 'src/assets/pages/es/stage-two.svg';
@@ -32,6 +32,10 @@ export function StageTwo() {
 
   const langContext = useLangContext();
   const StepImage = langContext.state.selected.startsWith('en') ? StepImageEN : StepImageES;
+
+  const onGoHomeHandler = () => {
+    navigate('/');
+  };
 
   const onBackHandler = () => {
     navigate('../stage-one');
@@ -79,6 +83,7 @@ export function StageTwo() {
     <>
       <Title type="page" label={t('Application Approval Process')} className="mb-4" />
       <Text className="text-lg mb-12">{application?.borrower.legal_name}</Text>
+      <Link to="../stage-two" />
       <img className="mb-14 ml-8" src={StepImage} alt="step" />
       <Title type="section" label={t('Stage 2: Business Documents')} className="mb-8" />
       <Text className="mb-4">{t('Review and verify the data for the business.')}</Text>
@@ -98,6 +103,9 @@ export function StageTwo() {
         <ApplicationDocumentsTable allowDataVerification className="xl:w-4/5" application={application} />
       )}
       <div className="mt-6 md:mb-8 grid grid-cols-1 gap-4 md:flex md:gap-0">
+        <div>
+          <Button className="md:mr-4" label={t('Go Home')} onClick={onGoHomeHandler} />
+        </div>
         <div>
           <Button className="md:mr-4" label={t('Go Back')} onClick={onBackHandler} />
         </div>
